+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.1.2] - 2026-09-03
 ### Fixed
 - fix(phpstan): drop unnecessary null coalesce operators in TemplateWhispererSuggestionForm
 - fix(phpstan): inject the entity type manager instead of storing the suggestion storage in TemplateWhispererManager
@@ -201,7 +203,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add BrowserTestBase to test web-based behaviors and interactions.
 - First draft.
 
-[Unreleased]: https://github.com/antistatique/drupal-template-whisperer/compare/4.1.1...HEAD
+[Unreleased]: https://github.com/antistatique/drupal-template-whisperer/compare/4.1.2...HEAD
+[4.1.2]: https://github.com/antistatique/drupal-template-whisperer/compare/4.1.1...4.1.2
 [4.1.1]: https://github.com/antistatique/drupal-template-whisperer/compare/4.1.0...4.1.1
 [4.1.0]: https://github.com/antistatique/drupal-template-whisperer/compare/4.0.6...4.1.0
 [4.0.6]: https://github.com/antistatique/drupal-template-whisperer/compare/4.0.4...4.0.6